+++ CHANGELOG.md
@@ -4,6 +4,7 @@ History of **rest2go** library with changes description.
 
 ## Next release
 
+- `FEATURE` Health check
 - `FEATURE` Filtering
 - `FEATURE` Pagination
 - `FEATURE` Database migrations by Goose